Mix Columns: Changed Multiplication to be Better
WARNING: Small rant incoming.
Previously the Mix Column multiplication used a very complex and hard to
understand method. This is because EVERY resource I looked at explained
it as such (b is the state values):
Take b and shift it left by one then check if it's larger than 256. If
it is xor it with this random value 0x1B. If you're just multiplying by
2 you're done other wise if you're multiplying by 3 xor this result with
b. Then if you're just multiplying by 3 you're done otherwise ...

This is NOT how the multiplication should be done for this project as it
is very hard to understand. In reality all that is happening is a
multiplication with the AES polynomial. The reason it's not explained
like this is because for most implementations you don't need to
understand why you're doing something just that something happens. This
results in people regurgitating the same information over and over until
no one knows where it can from. So I changed it to be just a regular
multiplication over GF(2^8) with the regular AES polynomial being very
frustrated that I had to waste days of my life trying to understand so
overly optomized method for multiplicating over GF(2^8). Anyways it's
fix now and hopefully people who use this as a resource won't be as
frustrated as I was.